# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: eBay Price Checker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /monitors:
    get:
      operationId: listMonitors
      tags: [monitors]
      summary: List all price monitors
      responses:
        "200":
          description: List of monitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Monitor"
    post:
      operationId: createMonitor
      tags: [monitors]
      summary: Create a new price monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonitorRequest"
      responses:
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /monitors/{id}:
    get:
      operationId: getMonitor
      tags: [monitors]
      summary: Get a specific monitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monitor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMonitor
      tags: [monitors]
      summary: Delete a monitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMonitor
      tags: [monitors]
      summary: Update monitor settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMonitorRequest"
      responses:
        "200":
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
  /monitors/{id}/check:
    post:
      operationId: checkMonitorNow
      tags: [monitors]
      summary: Manually trigger a price check for a monitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResult"
  /monitors/{id}/history:
    get:
      operationId: getMonitorHistory
      tags: [monitors]
      summary: Get price history for a monitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceHistory"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    CreateMonitorRequest:
      type: object
      properties:
        ebayUrl:
          type: string
        myPrice:
          type: number
        myCondition:
          type: string
        label:
          type: string
        spreadsheetRow:
          type: integer
          description: Row number in spreadsheet to update
        inventoryProductId:
          type: integer
        checkIntervalMinutes:
          type: integer
          default: 360
      required:
        - ebayUrl
        - myPrice
        - myCondition
    UpdateMonitorRequest:
      type: object
      properties:
        myPrice:
          type: number
        myCondition:
          type: string
        label:
          type: string
        spreadsheetRow:
          type: integer
        inventoryProductId:
          type: integer
          nullable: true
        checkIntervalMinutes:
          type: integer
        isActive:
          type: boolean
    Monitor:
      type: object
      properties:
        id:
          type: integer
        ebayUrl:
          type: string
        myPrice:
          type: number
        myCondition:
          type: string
        label:
          type: string
        spreadsheetRow:
          type: integer
        inventoryProductId:
          type: integer
        checkIntervalMinutes:
          type: integer
        isActive:
          type: boolean
        currentLowestPrice:
          type: number
        currentLowestCondition:
          type: string
        lastCheckedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        hasAlert:
          type: boolean
      required:
        - id
        - ebayUrl
        - myPrice
        - myCondition
        - isActive
        - createdAt
        - hasAlert
    MonitorDetail:
      allOf:
        - $ref: "#/components/schemas/Monitor"
        - type: object
          properties:
            recentHistory:
              type: array
              items:
                $ref: "#/components/schemas/PriceHistory"
            activeAlerts:
              type: array
              items:
                $ref: "#/components/schemas/Alert"
    CheckResult:
      type: object
      properties:
        monitorId:
          type: integer
        lowestPrice:
          type: number
        lowestCondition:
          type: string
        myPrice:
          type: number
        alertTriggered:
          type: boolean
        repricedTo:
          type: number
        evidenceUrls:
          type: array
          items:
            type: string
        message:
          type: string
        checkedAt:
          type: string
          format: date-time
      required:
        - monitorId
        - lowestPrice
        - myPrice
        - alertTriggered
        - evidenceUrls
        - checkedAt
    PriceHistory:
      type: object
      properties:
        id:
          type: integer
        monitorId:
          type: integer
        lowestPrice:
          type: number
        lowestCondition:
          type: string
        checkedAt:
          type: string
          format: date-time
      required:
        - id
        - monitorId
        - lowestPrice
        - checkedAt
    Alert:
      type: object
      properties:
        id:
          type: integer
        monitorId:
          type: integer
        monitorLabel:
          type: string
        competitorPrice:
          type: number
        competitorCondition:
          type: string
        myPrice:
          type: number
        priceDiff:
          type: number
        isResolved:
          type: boolean
        spreadsheetSynced:
          type: boolean
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required:
        - id
        - monitorId
        - competitorPrice
        - myPrice
        - priceDiff
        - isResolved
        - spreadsheetSynced
        - createdAt
...